# [bug#12284] Ensure ArrayBufferView is consistent with buffer