--- a/test/scalacheck/scala/collection/IteratorProperties.scala
+++ b/test/scalacheck/scala/collection/IteratorProperties.scala
@@ -29,17 +29,22 @@ object IteratorProperties extends Properties("Iterator") {
     case it: Iterator[Int] => View.dropRightIterator(it, n)
     case x                 => throw new MatchError(x)
   })
+  property("patch") = check((it, n) => it match {
+    case it: Iterable[Int] => it.iterator.patch(1, Iterator.empty, n)
+    case it: Iterator[Int] => it.patch(1, Iterator.empty, n)
+    case x                 => throw new MatchError(x)
+  })
 
   def check(f: (IterableOnceOps[Int, IterableOnce, IterableOnce[Int]], Int) => IterableOnce[Int]): Prop = forAll(Arbitrary.arbitrary[Seq[Int]], smallInteger) { (s: Seq[Int], n: Int) =>
     val indexed = s.toIndexedSeq // IndexedSeqs and their Iterators have a knownSize
     val simple = new SimpleIterable(s) // SimpleIterable and its Iterator don't
-    val stream = LazyList.from(s) // Lazy
+    val lazyList = LazyList.from(s) // Lazy
     val indexed1 = f(indexed, n).iterator.to(Seq)
     val indexed2 = f(indexed.iterator, n).iterator.to(Seq)
     val simple1 = f(simple, n).iterator.to(Seq)
     val simple2 = f(simple.iterator, n).iterator.to(Seq)
-    val stream1 = f(stream, n).iterator.to(Seq)
-    val stream2 = f(stream.iterator, n).iterator.to(Seq)
+    val stream1 = f(lazyList, n).iterator.to(Seq)
+    val stream2 = f(lazyList.iterator, n).iterator.to(Seq)
     (indexed1 == indexed2) :| s"indexed: $indexed1 != $indexed2" &&
       (simple1 == simple2) :| s"simple: $simple1 != $simple2" &&
       (stream1 == stream2) :| s"stream: $stream1 != $stream2" &&