Fixes scala/bug#12284

Based on top of #9258