@SethTisue 2021-08-12
@NthPortal hard to be totally sure, but afaict the `ArrayBufferBenchmark` I added over in #9258 should suffice for benchmarking this one as well