@NthPortal 2021-08-23
@SethTisue yes.

I'll have to review the changes, as it's been a really long time since I've looked at them, but I'm not even sure there's anything to benchmark