@SethTisue 2021-08-30
needs rebase on top of #9258, which is now merged