@NthPortal 2021-08-30
I'm benching `addAll` after changes at the moment. thankfully, running fewer benchmarks takes less time