@NthPortal 2021-08-31
I've written an Ammonite script to merge before and after JMH output files as a diff, if anyone wants it: https://gist.github.com/NthPortal/b21ea1d91f41a1c82a3113c0c3848da4

features:

- strip output prefix (usually `[info] ` in sbt)
- align the columns of the two outputs (most frequently misaligned in the "Error" column)
- deduplicates header when interleaving