@NthPortal 2021-08-31
benchmark results

```diff
-before
+after

 Benchmark                     (size)  Mode  Cnt      Score     Error  Units
-ArrayBufferBenchmark.addAll1      10  avgt   30    104.579 ±   3.018  ns/op
+ArrayBufferBenchmark.addAll1      10  avgt   30     99.248 ±   0.430  ns/op
-ArrayBufferBenchmark.addAll1     100  avgt   30    266.944 ±   0.580  ns/op
+ArrayBufferBenchmark.addAll1     100  avgt   30    266.565 ±   0.470  ns/op
-ArrayBufferBenchmark.addAll1    1000  avgt   30   1699.199 ±   7.797  ns/op
+ArrayBufferBenchmark.addAll1    1000  avgt   30   1697.700 ±   4.453  ns/op
-ArrayBufferBenchmark.addAll1   10000  avgt   30  24915.264 ±  50.435  ns/op
+ArrayBufferBenchmark.addAll1   10000  avgt   30  25152.311 ± 136.803  ns/op
-ArrayBufferBenchmark.addAll2      10  avgt   30    106.333 ±   2.565  ns/op
+ArrayBufferBenchmark.addAll2      10  avgt   30    119.342 ±   0.267  ns/op
-ArrayBufferBenchmark.addAll2     100  avgt   30    669.213 ±   5.293  ns/op
+ArrayBufferBenchmark.addAll2     100  avgt   30    784.811 ±   9.628  ns/op
-ArrayBufferBenchmark.addAll2    1000  avgt   30   7041.905 ±  13.658  ns/op
+ArrayBufferBenchmark.addAll2    1000  avgt   30   8561.646 ±  14.924  ns/op
-ArrayBufferBenchmark.addAll2   10000  avgt   30  81940.910 ± 152.333  ns/op
+ArrayBufferBenchmark.addAll2   10000  avgt   30  92149.025 ± 294.481  ns/op
-ArrayBufferBenchmark.addAll3      10  avgt   30    115.073 ±   2.267  ns/op
+ArrayBufferBenchmark.addAll3      10  avgt   30    134.425 ±   0.757  ns/op
-ArrayBufferBenchmark.addAll3     100  avgt   30    770.600 ±   2.999  ns/op
+ArrayBufferBenchmark.addAll3     100  avgt   30    761.728 ±   1.517  ns/op
-ArrayBufferBenchmark.addAll3    1000  avgt   30   6750.484 ±  11.538  ns/op
+ArrayBufferBenchmark.addAll3    1000  avgt   30   6688.084 ±  15.008  ns/op
-ArrayBufferBenchmark.addAll3   10000  avgt   30  78802.133 ± 137.187  ns/op
+ArrayBufferBenchmark.addAll3   10000  avgt   30  76560.228 ± 181.348  ns/op
```

sure doesn't look like an optimisation 😬