@NthPortal 2021-08-31
I re-ran the "after" benchmarks and got the same result.

I'm gonna remove the attempted optimisation and stick with just the cleanup