@NthPortal 2021-09-01
the new benchmark results are... confusing

```diff
 Benchmark                     (size)  Mode  Cnt       Score     Error  Units
-ArrayBufferBenchmark.addAll1      10  avgt   30     100.307 ±   0.263  ns/op
+ArrayBufferBenchmark.addAll1      10  avgt   30      99.412 ±   0.221  ns/op
-ArrayBufferBenchmark.addAll1     100  avgt   30     267.558 ±   0.700  ns/op
+ArrayBufferBenchmark.addAll1     100  avgt   30     267.451 ±   0.619  ns/op
-ArrayBufferBenchmark.addAll1    1000  avgt   30    1711.113 ±  16.689  ns/op
+ArrayBufferBenchmark.addAll1    1000  avgt   30    1686.831 ±   1.987  ns/op
-ArrayBufferBenchmark.addAll1   10000  avgt   30   25082.094 ± 153.114  ns/op
+ArrayBufferBenchmark.addAll1   10000  avgt   30   24900.748 ±  38.106  ns/op
-ArrayBufferBenchmark.addAll2      10  avgt   30     117.481 ±   0.499  ns/op
+ArrayBufferBenchmark.addAll2      10  avgt   30     117.389 ±   2.800  ns/op
-ArrayBufferBenchmark.addAll2     100  avgt   30    1083.291 ±   4.457  ns/op
+ArrayBufferBenchmark.addAll2     100  avgt   30     901.295 ±  13.703  ns/op
-ArrayBufferBenchmark.addAll2    1000  avgt   30    7396.998 ±  19.380  ns/op
+ArrayBufferBenchmark.addAll2    1000  avgt   30    8899.769 ± 230.113  ns/op
-ArrayBufferBenchmark.addAll2   10000  avgt   30  116549.685 ± 243.404  ns/op
+ArrayBufferBenchmark.addAll2   10000  avgt   30  103605.259 ± 226.673  ns/op
-ArrayBufferBenchmark.addAll3      10  avgt   30     103.075 ±   0.366  ns/op
+ArrayBufferBenchmark.addAll3      10  avgt   30     109.616 ±   3.312  ns/op
-ArrayBufferBenchmark.addAll3     100  avgt   30     666.471 ±   5.343  ns/op
+ArrayBufferBenchmark.addAll3     100  avgt   30     638.888 ±   1.148  ns/op
-ArrayBufferBenchmark.addAll3    1000  avgt   30    7072.978 ±  13.349  ns/op
+ArrayBufferBenchmark.addAll3    1000  avgt   30    7053.111 ±  14.073  ns/op
-ArrayBufferBenchmark.addAll3   10000  avgt   30   81894.343 ± 131.434  ns/op
+ArrayBufferBenchmark.addAll3   10000  avgt   30   82046.005 ± 181.125  ns/op
-ArrayBufferBenchmark.addAll4      10  avgt   30     114.865 ±   2.215  ns/op
+ArrayBufferBenchmark.addAll4      10  avgt   30     121.173 ±   4.157  ns/op
-ArrayBufferBenchmark.addAll4     100  avgt   30     781.175 ±   4.043  ns/op
+ArrayBufferBenchmark.addAll4     100  avgt   30     747.862 ±   2.178  ns/op
-ArrayBufferBenchmark.addAll4    1000  avgt   30    6749.267 ±  15.244  ns/op
+ArrayBufferBenchmark.addAll4    1000  avgt   30    6750.256 ±  10.804  ns/op
-ArrayBufferBenchmark.addAll4   10000  avgt   30   79043.508 ± 275.626  ns/op
+ArrayBufferBenchmark.addAll4   10000  avgt   30   78989.568 ± 132.373  ns/op
```

`addAll2` is the one that the new changes are meant to optimise, and the results are highly inconsistent. sometimes moderate improvements, sometimes moderate losses.

Either `ArrayBuffer#addAll(s.c.i.Set)` has really weird behaviour, or something caused slowdowns during benchmark execution. I'm gonna rerun `addAll2` with a 20 warmup and 40 rounds (rather than 15 and 15), to hopefully get more consistent results.