@NthPortal 2021-09-01
```diff
 Benchmark                     (size)  Mode  Cnt       Score     Error  Units
-ArrayBufferBenchmark.addAll2      10  avgt   80     120.388 ±   0.697  ns/op
+ArrayBufferBenchmark.addAll2      10  avgt   80     115.786 ±   0.649  ns/op
-ArrayBufferBenchmark.addAll2     100  avgt   80    1096.057 ±   2.217  ns/op
+ArrayBufferBenchmark.addAll2     100  avgt   80     941.184 ±   6.919  ns/op
-ArrayBufferBenchmark.addAll2    1000  avgt   80    7367.775 ±  16.810  ns/op
+ArrayBufferBenchmark.addAll2    1000  avgt   80    8626.615 ±  49.965  ns/op
-ArrayBufferBenchmark.addAll2   10000  avgt   80  103249.393 ± 517.055  ns/op
+ArrayBufferBenchmark.addAll2   10000  avgt   80  101780.067 ± 534.532  ns/op
```

... nope. I don't understand how, but nope. I'll revert the attempted optimisation completely, and just leave it at the tidying up and the `View.Patched` bugfix