@SethTisue 2021-09-02
Performance is a cruel and distant god.