@SethTisue 2021-09-15
2nd attempt to attract reviewer attention from @scala/collections crew and/or @julienrf