@SethTisue 2021-09-22
@NthPortal what's next here? are you interested in doing the ScalaCheck thing @johnynek suggested?