@NthPortal 2021-09-23
@SethTisue i've been and continue to be sick, but i'll try to add the scalacheck stuff today or tomorrow. if i don't get to it, and @johnynek has perms, they can feel free to add a commit to this branch with it